v1_db구축방식/requirements.txt
openai==1.0.0
python-dotenv==1.0.0
sqlite3